refactor: refactor api position
